tests: migrate unit tests to Rust
To avoid issues like python2 not available on distros, we should simply
migrate unit tests to Rust. That would make more sense, as the main code
is already written in Rust.

Adding edition in Cargo.toml.
Without edition = 2021, build fails due to a missing "extern crate ...".
Specifying an edition, Build passes without having to add it.

As Rust does not support global variables by default, it is necessary to
make use of lazy_static, so the variables could be accessed in multiple
tests.